Scoreboard support!!
We now have scoreboard support!
Added method for asynchronous sending of data to Score_Glen.php. This is achieved by hijacking the "submit" event of a new web form; and firing that event by dispatching a click event on a button in the form.

This is the only way I could find to make it work, while still using vanilla Javascript and not adding any more complexities.